specification/search/swagger_equiv/requirements.txt
# Core dependencies for semantic Swagger equivalency checking
PyYAML==6.0.1